break Node into its own file
