+++ src/core/cluster/state_machine.rs
@@ -21,15 +21,11 @@ use tokio::sync::RwLock as TokioRwLock;
 use uuid::Uuid;
 
 use super::{
-    errors::*,
-    handle::Response,
-    logs::CoyoteLogs,
-    raft::{Node, TypeConfig},
+    Node, NodeId, errors::*, handle::Response, logs::CoyoteLogs, raft::TypeConfig,
     serialized_state_machine,
 };
 use crate::AppState;
 
-type NodeId = <TypeConfig as RaftTypeConfig>::NodeId;
 type StorageError = openraft::StorageError<NodeId>;
 type StorageResult<T> = Result<T, StorageError>;
